Expose SvgIntrinsicDimensions::from_svg_source; fix svg_background_size test
